Crate future_bus

A bus type (SPMC) channel where multiple consumers can subscribe to single source channel. The bus can internally use any channel/receiver combination. By default it provides constructor methods to use the futures::channel::mpsc channels.

Structs

BusSubscriber
A receiver of data from a FutureBus. When it is dropped, the sender is removed from the bus. Note that this means dropping a receiver causes a write access to the bus, so it causes a lock.

FutureBus
The struct containing references to the receivers and at the same time the data source for the bus.

Functions

bounded
Create a new bus using bounded channels

unbounded
Create a new bus using unbounded channels